Rename reader_options.h to lcf_options.h This was probably the last trace of 'libreaders'...
