+++ builds/vs2015/liblcf.vcxproj.filters
@@ -433,6 +433,9 @@
     <ClInclude Include="..\..\src\inireader.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\lcf_options.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\src\ldb_reader.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -446,9 +449,6 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\reader_lcf.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\reader_options.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\reader_struct.h">
