+++ src/rpg_fixup.cpp
@@ -4,7 +4,7 @@
  * http://opensource.org/licenses/MIT
  */
 
-#include "reader_options.h"
+#include "lcf_options.h"
 #include "rpg_actor.h"
 #include "rpg_mapinfo.h"
 #include "rpg_system.h"
