+++ src/rpg_setup.cpp
@@ -4,7 +4,7 @@
  * http://opensource.org/licenses/MIT
  */
 
-#include "reader_options.h"
+#include "lcf_options.h"
 #include "rpg_actor.h"
 #include "rpg_event.h"
 #include "rpg_map.h"
