+++ src/writer_lcf.h
@@ -14,7 +14,7 @@
 #include <cstdio>
 #include <cassert>
 #include "reader_types.h"
-#include "reader_options.h"
+#include "lcf_options.h"
 #include "reader_util.h"
 
 /**
